# Create a faster script to populate reference caches.