@whitwham 2025-07-10
Your code is indeed faster.  Thank you very much.