@rhpvorderman 2025-07-11
Now you just have to wait for someone who implements it in rust ;-). 

Thanks for merging! Anyway, if people start complaining about unforeseen bugs, please ping me and I will fix them. I did do the due diligence thing with a test suite and testing various scenarious (larger than memory spool contigs, smaller than memory spool contigs etc.), but no test suite is perfect.